[mac] Fix #43, again (🤞)
My main strategy here was to create a minimal single-window objective-c
application (based on [1]) and then to trace our executation
and compare that against the execution of the objective-c app.

There were a few places where our behaviour differed, but changing
those had no apparent effect on application behaviour; but in
the course of that work I found the actual bug, which was much sneaker.

TL;DR: We were using [NSRunningApplication activateWithOptions:],
when we should have been using [NSApplication  activateIgnoringOtherApps:].
The NSRunningApplication API is intended for controlling other running
applications; we were likely sending the msg to nil.

[1]: https://www.cocoawithlove.com/2010/09/minimalist-cocoa-programming.html