[mac] Fix issue with menu not being interactive on mac
This is... fiddly.

This is related to de286f2; the fix implemented there overlooked an
issue where [NSApplication activateIgnoringOtherApps:] was being called
_before_ [NSApplication run:], which doesn't work; and without actually
figuring out how to get cocoa to use a custom class as our base
application class, it's easier to just use an app delegate and then call
activateIgnoring from there, which seems to work.

This feels a bit cludgy but I don't mind it too much because I expect we will
eventually want to have an AppDelegate anyway.

This solution was inspired on this SO question:
https://stackoverflow.com/questions/33345686/cocoa-application-menu-bar-not-clickable