druid-shell macOS text input
Adds APIs for cross-platform text input, and an initial implementation of text input
on macOS that allows for dead keys, IME input for Chinese, Japanese, etc.
On other platforms, the text input method calls are simulated from key_down events,
as a temporary measure until we can properly implement input methods on those platforms.

Also adds a new druid-shell example, `edit_text`, which is a rough, LTR and single-line
only implementation of the text input APIs to demonstrate that they work correctly.

A massive thanks to yvt's library tcw3 (https://github.com/yvt/Stella2), which was a fantastic
blueprint for the structure of this API, as well as indispensable reference when implementing
the macOS text input interfaces.