X11: Add support for transparent windows (#1803)
* X11: Add support for transparent windows

First, some history lesson:

X11 originally did not support transparency.  Visuals describe how pixel
values are interpreted, but they can only describe red/green/blue
components. Then, the RENDER extension came and added transparency. Now
we have picture formats. That's basically the same as a visual, but it
also describes an alpha component. There is a mapping between picture
formats and visual. Also, there is a standard ARGB32 format that is
always supported.

Then came the composite extension. Classically, the X11 server "drew
stuff" directly to the output, but there is no alpha compositing
involved.  With composite, a composite manager can request drawing to be
redirected to an off-screen pixmap. The composite manager will then
produce the "real" visible screen contents. And in doing so, it can take
alpha values into account.

This commit adds support for transparent windows to the X11 backend. For
that, it has to find the standard ARGB32 render picture format and the
corresponding visual. When a transparent window is created, it checks if
a composite manager is running. If not, nothing changes. If everything
worked out (RENDER is supported, we found the ARGB32 picture format, a
composite manager is running), transparent windows are created with this
visual instead of the root window's visual.

When a window has a different bit depth than its parent window, it has
to "fully define" its rendering. Since it has a different bit depth than
its parent, the X11 server cannot just copy background and window border
from the parent. Thus, when we create an ARGB32 window (which has
depth=32), we are creating a window with a different depth than the root
window (most likely depth=24) and have to provide some additional
settings. Even though no border will be used, we still have to define
its color.

Signed-off-by: Uli Schlachter <psychon@znc.in>

* Apply rustfmt's suggestion

Signed-off-by: Uli Schlachter <psychon@znc.in>

* Use the correct depths for pixmaps

The code always called Buffers::new() with a depth of screen.root_depth.
While this value was always correct previously, for transparency we are
using a window with depth=32. The mismatch in depths then caused
RENDER's create_picture to fail down the road.

For clarity, this commit also replaces a use of COPY_DEPTH_FROM_PARENT
with the correct depth. This change should not actually make any
difference in practice.

Signed-off-by: Uli Schlachter <psychon@znc.in>

* Extract helper function for extra readability

Signed-off-by: Uli Schlachter <psychon@znc.in>

* Use find_map() instead of filter_map() + next()

Signed-off-by: Uli Schlachter <psychon@znc.in>

* Fix cargo fmt

Signed-off-by: Uli Schlachter <psychon@znc.in>

* Another cargo fmt fix

Signed-off-by: Uli Schlachter <psychon@znc.in>

Co-authored-by: Manmeet Maan <49202620+Maan2003@users.noreply.github.com>