Initialize ns_window on mac
We were setting ns_view but not ns_window, which was causing the query to fetch scale to fail, and thus dpi wrong.

I'm not 100% sure we should be calling the `window` method to get this, we could store the NSWindow pointer in the WindowHandle. The main reason we didn't do this was to support the VST use case, but that's not actively being developed.

In any case, I think this is probably "good enough."