Add thread-safe app handle (#90)
Adds an AppHandle struct that can be shared across threads, as well as an associated Application:get_handle method to create instances of that struct. Similar to IdleHandle, AppHandle simply allows the user to schedule a function to be run on the main thread (at some point or potentially never).