Rename `DRUID_X11_DPI` and mention it in the docs. (#139)
It is now `GLAZIER_OVERRIDE_SCALE` in the event that other backends
decide to respect this value. It is also a scale, not a DPI, so the
name is changed accordingly.

This also removes a mention of an environment variable that wasn't
checked to control a behavior that no longer exists.